+++ changelog.md
@@ -1,9 +1,11 @@
-# dspic33a-code-examples v1.0.0
+# dspic33A Code Examples v1.0.0
 ### Release Highlights
 
+Initial Release of dsPIC33A code examples on GitHub.
 
 
 ### Features Added\Updated
+Added Code example for Real/Complex FFT using XC-DSC DSP Library.